+++ packaging/chocolatey/tools/chocolateyInstall.ps1
@@ -1,9 +1,9 @@
 $ErrorActionPreference = 'Stop'
 
 $packageName  = 'localpdf-studio'
-$version      = '0.15.0'
+$version      = '0.17.0'
 $url64        = "https://github.com/jeiel85/localpdf-studio/releases/download/v$version/LocalPDF.Studio_${version}_x64-setup.exe"
-$checksum64   = 'c76076008e979ce653d864b6549e7f813c3e1bfd40fa1320019671e3195fa6a5'
+$checksum64   = '02d3f33e7f5defa1adc8f5691485d2b381bc19d452d96400c3308bd0619c23b2'
 $checksumType = 'sha256'
 
 $packageArgs = @{